Created a modal file and transfered modal to that file
